requirements.txt
aioquic>=0.9.20
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.1
pytest>=7.3.1